pyproject.toml
[tool.poetry]
name = "rag-box"
version = "0.1.0"
description = ""
authors = ["Your Name <you@example.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.10"
fastapi = "^0.110.1"
python-dotenv = "^1.0.0"
llama-index = "0.10.37"
llama-index-core = "0.10.37"
llama-index-agent-openai = "0.2.5"
uvicorn = { version = "^0.23.2", extras = [ "standard" ] }
pydantic-settings = "^2.2.1"
python-multipart = "^0.0.9"
aiostream = "^0.5.2"
llama-index-vector-stores-chroma = "^0.1.8"
llama-index-embeddings-gemini = "^0.1.6"
llama-index-llms-gemini = "^0.1.7"
llama-index-llms-ollama = "^0.1.3"
llama-index-embeddings-ollama = "^0.1.2"
llama-index-tools-wikipedia = "^0.1.3"
llama-index-tools-duckduckgo = "^0.1.1"


[tool.poetry.group.dev.dependencies]
black = "^24.4.0"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
